+++ packages/anon-aadhaar-pcd/src/utils.ts
@@ -1,4 +1,9 @@
 import { SnarkJSProof, Proof } from './types'
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
+import { groth16 } from 'snarkjs'
+import { AnonAadhaarPCD } from '../src/pcd'
+import { BigNumberish } from '../src/types'
 
 const getSubstringIndex = (str: Buffer, substring: string, n: number) => {
   let times = 0
@@ -105,4 +110,94 @@ export function unpackProof(proof: Proof): SnarkJSProof {
     protocol: 'groth16',
     curve: 'bn128',
   }
+}
+
+/**
+ * Turn a groth16 proof into a call data format to use it as a transaction input.
+ * @param proof The proof compatible with Semaphore.
+ * @param _publicSignals Public signal from the proof.
+ * @returns {a, b, c, Input} which are the input needed to verify a proof in the Verifier smart contract.
+ */
+export async function exportCallDataGroth16(
+  proof: AnonAadhaarPCD['proof']['proof'],
+  _publicSignals: BigNumberish
+): Promise<{
+  a: [BigNumberish, BigNumberish]
+  b: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]]
+  c: [BigNumberish, BigNumberish]
+  Input: BigNumberish[]
+}> {
+  const input = [
+    ...splitToWords(BigInt(_publicSignals), BigInt(64), BigInt(32)),
+  ]
+
+  const calldata = await groth16.exportSolidityCallData(proof, input)
+
+  const argv = calldata
+    .replace(/["[\]\s]/g, '')
+    .split(',')
+    .map((x: string) => BigInt(x).toString())
+
+  const a: [BigNumberish, BigNumberish] = [argv[0], argv[1]]
+  const b: [[BigNumberish, BigNumberish], [BigNumberish, BigNumberish]] = [
+    [argv[2], argv[3]],
+    [argv[4], argv[5]],
+  ]
+  const c: [BigNumberish, BigNumberish] = [argv[6], argv[7]]
+  const Input = []
+
+  for (let i = 8; i < argv.length; i++) {
+    Input.push(argv[i])
+  }
+  return { a, b, c, Input }
+}
+
+import { subtle } from 'crypto'
+
+function buffToBigInt(buff: string): bigint {
+  return BigInt('0x' + Buffer.from(buff, 'base64url').toString('hex'))
+}
+
+async function generateRsaKey(hash = 'SHA-256') {
+  const publicExponent = new Uint8Array([1, 0, 1])
+  const modulusLength = 2048
+  const { publicKey, privateKey } = await subtle.generateKey(
+    {
+      name: 'RSASSA-PKCS1-v1_5',
+      modulusLength,
+      publicExponent,
+      hash,
+    },
+    true,
+    ['sign', 'verify']
+  )
+
+  return { publicKey, privateKey }
+}
+
+export async function genData(
+  data: string,
+  HASH_ALGO: string
+): Promise<[bigint, bigint, bigint, bigint]> {
+  const keys = await generateRsaKey(HASH_ALGO)
+
+  const public_key = await subtle.exportKey('jwk', keys.publicKey)
+
+  const enc = new TextEncoder()
+  const text = enc.encode(data)
+  const hash = BigInt(
+    '0x' + Buffer.from(await subtle.digest(HASH_ALGO, text)).toString('hex')
+  )
+
+  const sign_buff = await subtle.sign(
+    { name: 'RSASSA-PKCS1-v1_5', hash: HASH_ALGO },
+    keys.privateKey,
+    text
+  )
+
+  const e = buffToBigInt(public_key.e as string)
+  const n = buffToBigInt(public_key.n as string)
+  const sign = BigInt('0x' + Buffer.from(sign_buff).toString('hex'))
+
+  return [e, sign, n, hash]
 }